Shop 엔티티 서브 패키지 구분.
