+++ src/main/java/kyonggi/bookslyserver/domain/event/entity/ClosingEventMenu.java
@@ -1,7 +1,7 @@
 package kyonggi.bookslyserver.domain.event.entity;
 
 import jakarta.persistence.*;
-import kyonggi.bookslyserver.domain.shop.entity.Menu;
+import kyonggi.bookslyserver.domain.shop.entity.Menu.Menu;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
 import lombok.*;
