+++ src/main/java/kyonggi/bookslyserver/domain/event/entity/EmployeeTimeEvent.java
@@ -1,7 +1,7 @@
 package kyonggi.bookslyserver.domain.event.entity;
 
 import jakarta.persistence.*;
-import kyonggi.bookslyserver.domain.shop.entity.Employee;
+import kyonggi.bookslyserver.domain.shop.entity.Employee.Employee;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
 import lombok.*;
