+++ src/main/java/kyonggi/bookslyserver/domain/shop/entity/Employee/Employee.java
@@ -1,7 +1,8 @@
-package kyonggi.bookslyserver.domain.shop.entity;
+package kyonggi.bookslyserver.domain.shop.entity.Employee;
 
 import jakarta.persistence.*;
 import kyonggi.bookslyserver.domain.review.entity.Review;
+import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
 import lombok.*;
