+++ src/main/java/kyonggi/bookslyserver/domain/shop/entity/Menu/Menu.java
@@ -1,6 +1,7 @@
-package kyonggi.bookslyserver.domain.shop.entity;
+package kyonggi.bookslyserver.domain.shop.entity.Menu;
 
 import jakarta.persistence.*;
+import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
 import lombok.*;
