+++ src/main/java/kyonggi/bookslyserver/domain/shop/entity/Menu/MenuCategory.java
@@ -1,4 +1,4 @@
-package kyonggi.bookslyserver.domain.shop.entity;
+package kyonggi.bookslyserver.domain.shop.entity.Menu;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -6,16 +6,13 @@ import lombok.*;
 @Entity
 @Getter
 @Setter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class WorkDays {
-    // 근무날짜 테이블
+public class MenuCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="menu_category_id")
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "employee_id")
-    private Employee employee;
+    private String name;
 }
