--- src/main/java/kyonggi/bookslyserver/domain/shop/entity/MenuCategory.java
@@ -1,18 +1,0 @@
-package kyonggi.bookslyserver.domain.shop.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class MenuCategory {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="menu_category_id")
-    private Long id;
-
-    private String name;
-}
