+++ src/main/java/kyonggi/bookslyserver/domain/shop/entity/Shop/ShopImage.java
@@ -1,0 +1,29 @@
+package kyonggi.bookslyserver.domain.shop.entity.Shop;
+
+
+import jakarta.persistence.*;
+import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
+import lombok.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+
+public class ShopImage {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="shopimage_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="shop_id")
+    private Shop shop;
+
+
+    private String img_uri;
+
+    private boolean isRepresentative;
+}
