+++ src/main/java/kyonggi/bookslyserver/domain/user/entity/FavoriteShop.java
@@ -1,12 +1,9 @@
 package kyonggi.bookslyserver.domain.user.entity;
 
 import jakarta.persistence.*;
-import kyonggi.bookslyserver.domain.shop.entity.Shop;
-import kyonggi.bookslyserver.domain.user.entity.User;
+import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
 import lombok.*;
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
 
 import static jakarta.persistence.FetchType.LAZY;
